+++ Changelog.md
@@ -6,6 +6,8 @@
     * Contribution from Martin Pecka
     * [Pull request 153](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/153)
 
+### Ignition Math 3.0.0 (2017-01-05)
+
 1. Deprecate many IGN_* macros in favor of static const variables in Helpers.hh
     * [Pull request 138](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/138)
     * [Pull request 137](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/137)